Extract renderer runtime from react-cosmos package
