--- packages/react-cosmos-fixture/index.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/index';
