--- packages/react-cosmos-fixture/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/index');
