+++ packages/react-cosmos-fixture/src/dom/ErrorCatch.tsx
