+++ packages/react-cosmos-fixture/src/dom/globalErrorHandler.ts
