+++ packages/react-cosmos-fixture/src/dom/rendererConnect.tsx
