+++ packages/react-cosmos-fixture/src/dom/shared.tsx
