+++ packages/react-cosmos/src/plugins/userDepsFile.tsx
@@ -1,17 +1,17 @@
-import path from 'path';
+import { FSWatcher, watch } from 'chokidar';
 import { writeFile } from 'fs';
-import promisify from 'util.promisify';
-import { watch, FSWatcher } from 'chokidar';
 import { debounce } from 'lodash';
+import path from 'path';
+import { NativeRendererConfig } from 'react-cosmos-shared2/renderer';
+import promisify from 'util.promisify';
 import { CosmosConfig } from '../config';
 import { DevServerPluginArgs } from '../shared/devServer';
 import {
-  getFixturePatterns,
+  generateUserDepsModule,
   getDecoratorPatterns,
-  getIgnorePatterns,
-  generateUserDepsModule
+  getFixturePatterns,
+  getIgnorePatterns
 } from '../shared/userDeps';
-import { NativeRendererConfig } from '../shared/rendererConfig';
 
 const writeFileAsync = promisify(writeFile);
