+++ packages/react-cosmos/src/plugins/webpack/client/reactDevtoolsHook.ts
