+++ packages/react-cosmos/src/plugins/webpack/client/userDeps.ts
@@ -1,8 +1,8 @@
 import {
-  ReactFixtureExportsByPath,
-  ReactDecoratorsByPath
+  ReactDecoratorsByPath,
+  ReactFixtureExportsByPath
 } from 'react-cosmos-shared2/react';
-import { DomRendererConfig } from '../../../shared/rendererConfig';
+import { DomRendererConfig } from 'react-cosmos-shared2/renderer';
 
 // NOTE: Renderer data is statically injected at compile time
 export const rendererConfig: DomRendererConfig = {
