+++ packages/react-cosmos/src/plugins/webpack/webpackConfig/__tests__/customDevConfig.ts
@@ -80,9 +80,7 @@ it('includes client entry', async () => {
 
 it('includes DOM devtooks hook entry', async () => {
   const { entry } = await getCustomDevWebpackConfig();
-  expect(entry).toContain(
-    require.resolve('../../../../domRenderer/reactDevtoolsHook')
-  );
+  expect(entry).toContain(require.resolve('../../client/reactDevtoolsHook'));
 });
 
 it('includes webpack-hot-middleware entry', async () => {
