+++ packages/react-cosmos/src/plugins/webpack/webpackConfig/__tests__/customExportConfig.ts
@@ -52,9 +52,7 @@ it('includes client entry', async () => {
 
 it('includes DOM devtooks hook entry', async () => {
   const { entry } = await getCustomExportWebpackConfig();
-  expect(entry).toContain(
-    require.resolve('../../../../domRenderer/reactDevtoolsHook')
-  );
+  expect(entry).toContain(require.resolve('../../client/reactDevtoolsHook'));
 });
 
 it('does not include webpack-hot-middleware entry', async () => {
