--- packages/react-cosmos/src/shared/rendererConfig.ts
@@ -1,7 +1,0 @@
-export type DomRendererConfig = {
-  containerQuerySelector: null | string;
-};
-
-export type NativeRendererConfig = {
-  port: number;
-};
